requirements.txt
uliweb-alembic
docker==7.1.0
fastapi==0.121.1
pydantic==2.12.4
python-dotenv==1.2.1
SQLAlchemy==2.0.44
uvicorn==0.38.0
psycopg[binary]==3.2.12
alembic==1.13.2
pytest==9.0.1
pytest-asyncio==1.3.0
httpx==0.28.1
